% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_blocks}
\alias{get_blocks}
\title{Get blocks}
\usage{
get_blocks(
  num_items = 24,
  melody_length = 3:15,
  instrument = c("Violin", "Cello")
)
}
\arguments{
\item{melody_length}{}
}
\description{
Get blocks
}
