% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_items}
\alias{get_items}
\title{Get items}
\usage{
get_items(
  num_items = 12,
  melody_length = 3:15,
  instrument = c("Violin", "Cello")
)
}
\arguments{
\item{instrument}{}
}
\description{
Get items
}
