% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_battery.R
\name{run_battery}
\alias{run_battery}
\title{Run the Suzuki Battery}
\usage{
run_battery(
  title = "Playing by Ear",
  app_name = "pbetsuzuki2024",
  instrument = c("Violin", "Cello"),
  musicassessr_aws = FALSE
)
}
\arguments{
\item{musicassessr_aws}{}
}
\description{
Run the Suzuki Battery
}
