% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suzuki_selected_item_bank_cello}
\alias{suzuki_selected_item_bank_cello}
\title{The Suzuki DB
#'}
\format{
Dataframe.
}
\usage{
suzuki_selected_item_bank_cello
}
\description{
The Suzuki DB
#'
}
\keyword{datasets}
